Integracion de servicion

+++ src/main/java/com/uptc/frw/jpa/repository/ReferenciaRepository.java
@@ -3,6 +3,6 @@ package com.uptc.frw.jpa.repository;
 import com.uptc.frw.jpa.model.Referencia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReferenciaRepository extends JpaRepository<Referencia, Long> {
+public interface ReferenciaRepository extends JpaRepository<Referencia, Integer> {
 }
